requirements.txt
requests
feedparser
jinja2
weasyprint
python-dotenv
